gl-aip/examples/multi-agent/pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Multi-agent coordinator pattern"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.0",
    "glaip-sdk",
    "gllm-core-binary>=0.3.0",
    "gllm-plugin-binary>=0.0.7",
    "gllm-tools-binary>=0.1.3",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
